Update header logging configuration and remove router
